fix type of query by method name parameter
there is no requirement in the spec that providers convert
long to int in repository method parameters.

+++ tck/src/main/java/ee/jakarta/tck/data/framework/read/only/AsciiCharacters.java
@@ -76,7 +76,7 @@ public interface AsciiCharacters extends DataRepository<AsciiCharacter, Long>, I
 
     List<AsciiCharacter> findByNumericValueLessThanEqualAndNumericValueGreaterThanEqual(int max, int min);
 
-    AsciiCharacter[] findFirst3ByNumericValueGreaterThanEqualAndHexadecimalEndsWith(long minValue, String lastHexDigit, Sort<AsciiCharacter> sort);
+    AsciiCharacter[] findFirst3ByNumericValueGreaterThanEqualAndHexadecimalEndsWith(int minValue, String lastHexDigit, Sort<AsciiCharacter> sort);
 
     Optional<AsciiCharacter> findFirstByHexadecimalStartsWithAndIsControlOrderByIdAsc(String firstHexDigit, boolean isControlChar);
